Add EVAL text enricher
