+++ module/projectfu.mjs
@@ -43,16 +43,16 @@ import { FUOptionalFeatureSheet } from './sheets/item-optional-feature-sheet.mjs
 import { OptionalFeatureDataModel, RollableOptionalFeatureDataModel } from './documents/items/optionalFeature/optional-feature-data-model.mjs';
 import { registerOptionalFeatures } from './documents/items/optionalFeature/optional-features.mjs';
 
-import { rolldataHtmlEnricher } from './helpers/rolldata-html-enricher.mjs';
+import { rolldataHtmlEnricher } from './enrichers/rolldata-html-enricher.mjs';
 import { FUActiveEffect } from './documents/effects/active-effect.mjs';
-import { InlineDamage } from './helpers/inline-damage.mjs';
+import { InlineDamage } from './enrichers/inline-damage.mjs';
 import { CanvasDragDrop } from './helpers/canvas-drag-drop.mjs';
-import { InlineResources } from './helpers/inline-resources.mjs';
-import { InlineChecks } from './helpers/inline-check.mjs';
+import { InlineResources } from './enrichers/inline-resources.mjs';
+import { InlineChecks } from './enrichers/inline-check.mjs';
 import { Flags } from './helpers/flags.mjs';
-import { InlineIcon } from './helpers/inline-icons.mjs';
+import { InlineCommon } from './enrichers/inline-common.mjs';
 import { TextEditorCommandDropdown } from './helpers/text-editor-command-dropdown.mjs';
-import { InlineEffects } from './helpers/inline-effects.mjs';
+import { InlineEffects } from './enrichers/inline-effects.mjs';
 import { SystemControls } from './helpers/system-controls.mjs';
 import { PlayerListEnhancements } from './helpers/player-list-enhancements.mjs';
 import { Checks } from './checks/checks.mjs';
@@ -64,14 +64,14 @@ import { ItemCustomizer } from './helpers/item-customizer.mjs';
 import { FUHooks } from './hooks.mjs';
 import { DamagePipeline } from './pipelines/damage-pipeline.mjs';
 import { ResourcePipeline } from './pipelines/resource-pipeline.mjs';
-import { InlineWeapon } from './helpers/inline-weapon.mjs';
+import { InlineWeapon } from './enrichers/inline-weapon.mjs';
 import { InlineHelper } from './helpers/inline-helper.mjs';
 import { Effects } from './pipelines/effects.mjs';
-import { InlineType } from './helpers/inline-type.mjs';
+import { InlineType } from './enrichers/inline-type.mjs';
 import { InvokerIntegration } from './documents/items/classFeature/invoker/invoker-integration.mjs';
 import { FUActiveEffectModel } from './documents/effects/active-effect-model.mjs';
 import { FUActiveEffectConfig } from './documents/effects/active-effect-config.mjs';
-import { InlineClocks } from './helpers/inline-clocks.mjs';
+import { InlineClocks } from './enrichers/inline-clocks.mjs';
 import { PartyDataModel } from './documents/actors/party/party-data-model.mjs';
 import { FUPartySheet } from './sheets/actor-party-sheet.mjs';
 import { StashDataModel } from './documents/actors/stash/stash-data-model.mjs';
@@ -105,7 +105,7 @@ import { RuleTriggerRegistry } from './documents/effects/triggers/rule-trigger-d
 import { RulePredicateRegistry } from './documents/effects/predicates/rule-predicate-data-model.mjs';
 import { ProgressPipeline } from './pipelines/progress-pipeline.mjs';
 import { ApplicationPipeline } from './pipelines/application-pipeline.mjs';
-import { InlineAction } from './helpers/inline-action.mjs';
+import { InlineAction } from './enrichers/inline-action.mjs';
 import { CompendiumBrowser } from './ui/compendium/compendium-browser.mjs';
 import { CompendiumIndex } from './ui/compendium/compendium-index.mjs';
 import { PressureSystem } from './systems/pressure-system.mjs';
@@ -399,7 +399,7 @@ Hooks.once('init', async () => {
 	InlineHelper.registerCommand(InlineWeapon);
 	InlineHelper.registerCommand(InlineType);
 	InlineHelper.registerCommand(InlineClocks);
-	InlineHelper.registerCommand(InlineIcon);
+	InlineHelper.registerCommand(InlineCommon);
 	InlineHelper.registerCommand(InlineAction);
 
 	Hooks.on('dropCanvasData', CanvasDragDrop.onDropCanvasData);
