+++ styles/css/global/inline.css
@@ -102,7 +102,7 @@ a.inline-roll {
 	box-shadow: 0.063rem 0.1rem var(--pfu-color-app-control-shadow);
 	border: 0.063rem solid var(--pfu-color-app-control-border);
 	text-shadow: 1px 1px var(--pfu-color-app-control-content-shadow);
-	margin: unset;
+	margin: 0 0 1px 0;
 	padding: 0 0.25rem;
 	display: inline-flex;
 	vertical-align: text-top;
